pyproject.toml
[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[project]
name = "minecord"
version = "0.1.0"
description = "A Discord bot for managing a Minecraft server."
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.0.0",
    "python-dotenv",
    "mcrcon",
    "PyYAML>=6.0",
]

[project.scripts]
minecord = "minecord.bot:run_bot"

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
]

[tool.hatch.build.targets.wheel]
# Explicitly tell Hatch where to find the package
packages = ["minecord"]